DI Factory object has been created.
